refactor: split preprocessor directives
